Fixed major problem with namespace and class collisions. Added ErrorResult helper methods.
