+++ DotNetElements.Core.ResultObject.Examples/DotNetElements.Core.ResultObject.Examples.csproj
