+++ DotNetElements.Core.ResultObject.Examples/ExampleErrorResult.cs
@@ -1,4 +1,4 @@
-﻿namespace DotNetElements.Core.Result.Examples;
+﻿namespace DotNetElements.Core.ResultObject.Examples;
 
 [ErrorResult<ExampleError>]
 public partial class ExampleErrorResult<TValue>;
