+++ DotNetElements.Core.ResultObject.Examples/Program.cs
@@ -1,7 +1,7 @@
-﻿using DotNetElements.Core.Result;
-using DotNetElements.Core.Result.Examples;
+﻿using DotNetElements.Core.ResultObject;
+using DotNetElements.Core.ResultObject.Examples;
 
-using static DotNetElements.Core.Result.Examples.ExampleErrorResultHelper;
+using static DotNetElements.Core.ResultObject.Examples.ExampleErrorResultHelper;
 
 ExampleErrorResult<string> result = GetExampleResult();
