+++ DotNetElements.Core.ResultObject.SourceGenerator/DotNetElements.Core.ResultObject.SourceGenerator.csproj
