+++ DotNetElements.Core.ResultObject.Tests/ErrorResultTests.cs
@@ -1,0 +1,10 @@
+﻿namespace DotNetElements.Core.ResultObject.Tests;
+
+[ErrorResult<string>]
+public partial class TestErrorResult;
+
+[TestClass]
+public class ErrorResultTests
+{
+    // todo
+}
