+++ DotNetElements.Core.ResultObject/ErrorResultAttribute.cs
@@ -1,4 +1,4 @@
-﻿namespace DotNetElements.Core.Result;
+﻿namespace DotNetElements.Core.ResultObject;
 
 [System.AttributeUsage(System.AttributeTargets.Class, AllowMultiple = false, Inherited = false)]
 public sealed class ErrorResultAttribute<TError> : System.Attribute;
