+++ DotNetElements.Core.ResultObject/Polyfill/IsExternalInit.cs
